requirements.txt
openai
streamlit
PyPDF2
langchain
python-dotenv
faiss-cpu
huggingface_hub
langchain_community
PyMuPDF
spaCy
nltk
pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
matplotlib
bitsandbytes 
accelerates


-e .